requirements.txt

python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
gradio>=4.0.0
requests>=2.31.0
httpx>=0.25.0
pydantic>=2.4.0
python-multipart>=0.0.6
aiofiles>=23.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
shap>=0.43.0
lime>=0.2.0.1
hopsworks>=3.4.0
hsml>=3.4.0
pyarrow>=14.0.0
confluent-kafka>=2.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
tqdm>=4.66.0
statsmodels